BreadCrum, Validation, Authorize, Fixed error From console,

+++ InventoryManagement/wwwroot/js/site.js
@@ -21,9 +21,11 @@ $(document).ready(function () {
                 "data": "roleId", "orderable": false, "render": function (data, type, row) {
 
                     var deleteUrl = "/Roles/Delete/" + row.roleId;
-
-                    return "<a href='/Roles/Edit/" + row.roleId + "'  class='btn btn-primary btn-sm' style='margin-left:5px' ><i class='far fa-edit'></i> Edit</a><a href='#' class='btn btn-danger btn-sm' style='margin-left:5px' onclick=deleteConfirm('" + deleteUrl + "'); ><i class='far fa-trash-alt'></i> Delete</a>";
-
+                    if (row.roleName != "SuperAdmin" && row.roleName != "User") {
+                        return "<a href='/Roles/Edit/" + row.roleId + "'  class='btn btn-primary btn-sm' style='margin-left:5px' ><i class='far fa-edit'></i> Edit</a><a href='#' class='btn btn-danger btn-sm' style='margin-left:5px' onclick=deleteConfirm('" + deleteUrl + "'); ><i class='far fa-trash-alt'></i> Delete</a>";
+                    } else {
+                        return "<a href='#' style='display:none' class='btn btn-primary btn-sm'><i class='far fa-edit'></i> Edit</a>";
+                    }
                     //} 
 
                 }
@@ -124,7 +126,7 @@ $(document).ready(function () {
             },
             {
                 "data": "listItemId", "orderable": false, "render": function (data, type, row) {
-                    var deleteUrl = "/Employee/Delete/" + row.employeeId;
+                    var deleteUrl = "/Employee/DeleteEmployee/" + row.employeeId;
 
                     return "<a href='/Employee/EditEmployee/" + row.employeeId + "'  class='btn btn-primary btn-sm' style='margin-left:5px' ><i class='far fa-edit'></i> Edit</a><a href='#' class='btn btn-danger btn-sm' style='margin-left:5px' onclick=deleteConfirm('" + deleteUrl + "'); ><i class='far fa-trash-alt'></i> Delete</a>";
 
@@ -161,6 +163,7 @@ function deleteData() {
                 $('#categoryDataTable').DataTable().ajax.reload();
                 $('#roleDatatable').DataTable().ajax.reload();
                 $('#itemDataTable').DataTable().ajax.reload();
+                $('#employeeDataTable').DataTable().ajax.reload();
 
                 toastr.success(data.message, 'Success Alert', { timeout: 300 });
             } else {
